+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+next version
+------------
+
+* Add modeling.drop_columns
+* Add modeling.drop_dropped_columns
+
 0.46.0 (2022-01-27)
 -------------------